requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
reportlab>=4.0.0
Werkzeug>=3.0.0
python-dotenv>=1.0.1
openai>=1.50.0
httpx>=0.27.2
requests>=2.31.0
PyMySQL>=1.1.0
psycopg2-binary>=2.9.9
authlib>=1.3.0
Flask-Login>=0.6.3
Flask-Limiter>=3.5.0
email-validator>=2.1.0
Flask-SocketIO>=5.3.6
python-socketio>=5.10.0
better-profanity>=0.7.0
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
python-json-logger>=2.0.7
PyJWT>=2.8.0
cryptography>=41.0.0
mediapipe>=0.10.0
gunicorn>=21.2.0
eventlet>=0.33.3